load more global data

+++ src/data_3d.cpp
@@ -94,7 +94,7 @@ namespace bl {
         }
         auto [my, _] = pos_.get_y_range(this->version_);
         y += my;
-        if (y >= this->biomes_.size() || y < 0) {
+        if (y >= static_cast<int>(this->biomes_.size()) || y < 0) {
             return biome::none;
         }
         return this->biomes_[y][cx][cz];
